+++ src/main/java/UDD/AleksaColovic/SearchEngine/service/LawService.java
@@ -6,6 +6,7 @@ import UDD.AleksaColovic.SearchEngine.repository.LawRepository;
 import UDD.AleksaColovic.SearchEngine.service.common.MinioService;
 import UDD.AleksaColovic.SearchEngine.service.helpers.SearchHelper;
 import UDD.AleksaColovic.SearchEngine.service.interfaces.ISearchService;
+import co.elastic.clients.elasticsearch._types.GeoLocation;
 import co.elastic.clients.elasticsearch._types.query_dsl.Query;
 import lombok.RequiredArgsConstructor;
 import org.apache.pdfbox.pdmodel.PDDocument;
@@ -17,35 +18,38 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.OptionalDouble;
 import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 public class LawService implements ISearchService<LawDocument> {
+    //region: Fields
     private final LawRepository lawRepository;
     private final MinioService minioService;
+    private final String bucketName = "laws";
 
     private final SearchHelper<LawDocument> searchHelper;
-
-    @Override
-    public void create(LawDocument document) {
-        lawRepository.save(document);
-    }
+    //endregion
 
     @Override
     public void upload(MultipartFile file) throws Exception {
-        if (minioService.loadFile(file.getOriginalFilename()) != null) {
-            throw new Exception(String.format("File with the given name already exists %s.", file.getOriginalFilename()));
+        if (minioService.checkIfExists(file.getOriginalFilename(), bucketName)) {
+            throw new Exception(String.format("Law file with the given file name: [%s] already exists.", file.getOriginalFilename()));
+        }
+        if (lawRepository.findByFileName(file.getOriginalFilename()) != null){
+            throw new Exception(String.format("Law index with the given file name: [%s] already exists.", file.getOriginalFilename()));
         }
 
-        minioService.uploadFile(file.getOriginalFilename(), file);
+        minioService.uploadFile(file.getOriginalFilename(), file, bucketName);
 
         PDDocument document = PDDocument.load(file.getInputStream());
         PDFTextStripper pdfStripper = new PDFTextStripper();
         String text = pdfStripper.getText(document);
         document.close();
 
-        create(new LawDocument(UUID.randomUUID(), text, file.getOriginalFilename()));
+        lawRepository.save(new LawDocument(UUID.randomUUID(), text, file.getOriginalFilename()));
     }
 
     @Override
@@ -59,7 +63,7 @@ public class LawService implements ISearchService<LawDocument> {
     }
 
     @Override
-    public List<SearchHit<LawDocument>> search(List<SearchItem> searchItems) {
+    public List<SearchHit<LawDocument>> search(List<SearchItem> searchItems, OptionalDouble radius) {
         Query searchQuery = searchHelper.buildSearchQuery(searchItems);
 
         NativeQuery nativeQuery = searchHelper.buildNativeQuery(searchQuery);
@@ -70,7 +74,16 @@ public class LawService implements ISearchService<LawDocument> {
     }
 
     @Override
-    public void delete(UUID id) {
+    public void delete(UUID id) throws Exception {
+        Optional<LawDocument> document = lawRepository.findById(id);
+        if (document.isEmpty()){
+            throw new Exception(String.format("Law index with the given id: [%s] does not exist.", id));
+        }
+        if (!minioService.checkIfExists(document.get().getFileName(), bucketName)) {
+            throw new Exception(String.format("Law file with the given id: [%s] does not exist.", id));
+        }
+
+        minioService.deleteFile(document.get().getFileName(), bucketName);
         lawRepository.deleteById(id);
     }
 }
